+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.4.4] - 2014-04-19
 ### Added
 - added `NewWhere` and `WithWhere` (#35)
+- added `NewOrderBy` and `WithOrderBy` (#36)
+- added `LimitResult` and `LimitOption` (#36)
 
 ## [1.4.3] - 2014-04-12
 ### Fixes